chore: rename files with correct case
