+++ lib/api/readFile.js
