+++ lib/api/rename.js
@@ -30,14 +30,14 @@ module.exports = function rename(oldPath, newPath, cb) {
         connection,
         function(err, file) {
           if (err) cb && cb(err);
-          else rename(connection, file, newPath, cb);
+          else rename_(connection, file, newPath, cb);
         }
       );
-    else rename(connection, file, newPath, cb);
+    else rename_(connection, file, newPath, cb);
   });
 };
 
-function rename(connection, file, newPath, cb) {
+function rename_(connection, file, newPath, cb) {
   // SMB2 rename
   SMB2Request(
     'set_info',
