+++ lib/api/writeFile.js
