+++ lib/smb2.js
@@ -90,7 +90,7 @@ proto.rename = autoPromise(
 );
 
 proto.readFile = autoPromise(
-  SMB2Connection.requireConnect(require('./api/readfile'))
+  SMB2Connection.requireConnect(require('./api/readFile'))
 );
 proto.createReadStream = autoPromise(
   SMB2Connection.requireConnect(require('./api/createReadStream'))
@@ -99,7 +99,7 @@ proto.createWriteStream = autoPromise(
   SMB2Connection.requireConnect(require('./api/createWriteStream'))
 );
 proto.writeFile = autoPromise(
-  SMB2Connection.requireConnect(require('./api/writefile'))
+  SMB2Connection.requireConnect(require('./api/writeFile'))
 );
 proto.unlink = autoPromise(
   SMB2Connection.requireConnect(require('./api/unlink'))
